refactor : 점심 조회 로직 수정, 점심 투표수 컬럼 추가에 따른 리팩토링
* refactor : 점심 메뉴 예외처리 빈 배열 응답으로 변경, pollCount 컬럼 추가

* refactor : pollCount 필드 추가에 따른 투표 로직 변경

- 연관관계 편의 메소드에 pollCount 변경 로직 추가

* refactor : 로직 변경, 컬럼 추가에 따른 테스트 코드 리팩토링

+++ src/main/java/com/ssafy/ssafsound/domain/lunch/dto/PostLunchPollResDto.java
@@ -7,9 +7,9 @@ import lombok.Getter;
 @Builder
 public class PostLunchPollResDto {
 
-    private Long pollCount;
+    private Integer pollCount;
 
-    public static PostLunchPollResDto of(Long pollCount){
+    public static PostLunchPollResDto of(Integer pollCount){
         return PostLunchPollResDto.builder()
                 .pollCount(pollCount)
                 .build();
